requirements.txt
fastapi
mangum
langgraph
langchain-anthropic
langsmith
weaviate-client
supabase
pydantic
pydantic-settings
pyyaml
pygithub
pytest
pytest-asyncio
httpx
uvicorn
python-dotenv
